Move child_document_archival from EspooConfig to EvakaEnv
